requirements.txt
# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Streamlit for web interface
streamlit>=1.28.0

# Timezone handling
pytz>=2023.3

# Yahoo Finance data
yfinance>=0.2.31

# Data manipulation and display
pandas>=2.0.0
tabulate>=0.9.0

